moved spark api to __init__
